Added support for pushstate
For pushstate to work properly, the server must return the content
of your main HTML page for arbitrary paths, as otherwise when a
user tries to reload a pushstate-enabled web app they would receive
a 404. This is now possible inside of slinky, by specifying a pushstate
hash in your slinky.yaml config file. The syntax is essentially the same
as defining proxies, and allows you to have different pushstate indices
for different paths. E.g.:

  pushstate:
    "/app1": "/index.haml"
    "/app2": "/index2.haml"

With this config, all non-existant paths beginning with /app1 will be
returned with index.haml and likewise with /app2 and index2.haml. It
is of course possible also to specify a blanket rule with "/". In the
case of conflicting rules, the more specific wins.